+++ sdk/data/aztables/client.go
@@ -125,6 +125,9 @@ func (t *Client) CreateTable(ctx context.Context, options *CreateTableOptions) (
 		options = &CreateTableOptions{}
 	}
 	resp, err := t.client.Create(ctx, generated.Enum1Three0, generated.TableProperties{TableName: &t.name}, options.toGenerated(), &generated.QueryOptions{})
+	if err != nil {
+		return CreateTableResponse{}, err
+	}
 	return createTableResponseFromGen(&resp), err
 }
 
@@ -373,6 +376,9 @@ func (t *Client) DeleteEntity(ctx context.Context, partitionKey string, rowKey s
 		options.IfMatch = &nilEtag
 	}
 	resp, err := t.client.DeleteEntity(ctx, generated.Enum1Three0, t.name, partitionKey, rowKey, string(*options.IfMatch), options.toGenerated(), &generated.QueryOptions{})
+	if err != nil {
+		return DeleteEntityResponse{}, err
+	}
 	return deleteEntityResponseFromGenerated(&resp), err
 }
 
@@ -476,6 +482,9 @@ func (t *Client) UpdateEntity(ctx context.Context, entity []byte, options *Updat
 			options.toGeneratedMergeEntity(mapEntity),
 			&generated.QueryOptions{},
 		)
+		if err != nil {
+			return UpdateEntityResponse{}, err
+		}
 		return updateEntityResponseFromMergeGenerated(&resp), err
 	case UpdateModeReplace:
 		resp, err := t.client.UpdateEntity(
@@ -487,6 +496,9 @@ func (t *Client) UpdateEntity(ctx context.Context, entity []byte, options *Updat
 			options.toGeneratedUpdateEntity(mapEntity),
 			&generated.QueryOptions{},
 		)
+		if err != nil {
+			return UpdateEntityResponse{}, err
+		}
 		return updateEntityResponseFromUpdateGenerated(&resp), err
 	}
 	if pk == "" || rk == "" {
@@ -495,8 +507,8 @@ func (t *Client) UpdateEntity(ctx context.Context, entity []byte, options *Updat
 	return UpdateEntityResponse{}, errInvalidUpdateMode
 }
 
-// InsertEntityOptions contains optional parameters for Client.InsertEntity
-type InsertEntityOptions struct {
+// UpsertEntityOptions contains optional parameters for Client.InsertEntity
+type UpsertEntityOptions struct {
 	// ETag is the optional etag for the Table
 	ETag azcore.ETag
 
@@ -505,35 +517,35 @@ type InsertEntityOptions struct {
 	UpdateMode UpdateMode
 }
 
-// InsertEntityResponse contains response fields for Client.InsertEntity
-type InsertEntityResponse struct {
+// UpsertEntityResponse contains response fields for Client.InsertEntity
+type UpsertEntityResponse struct {
 	ETag azcore.ETag
 }
 
-func insertEntityFromGeneratedMerge(g *generated.TableClientMergeEntityResponse) InsertEntityResponse {
+func insertEntityFromGeneratedMerge(g *generated.TableClientMergeEntityResponse) UpsertEntityResponse {
 	if g == nil {
-		return InsertEntityResponse{}
+		return UpsertEntityResponse{}
 	}
 
 	var ETag azcore.ETag
 	if g.ETag != nil {
 		ETag = azcore.ETag(*g.ETag)
 	}
-	return InsertEntityResponse{
+	return UpsertEntityResponse{
 		ETag: ETag,
 	}
 }
 
-func insertEntityFromGeneratedUpdate(g *generated.TableClientUpdateEntityResponse) InsertEntityResponse {
+func insertEntityFromGeneratedUpdate(g *generated.TableClientUpdateEntityResponse) UpsertEntityResponse {
 	if g == nil {
-		return InsertEntityResponse{}
+		return UpsertEntityResponse{}
 	}
 
 	var ETag azcore.ETag
 	if g.ETag != nil {
 		ETag = azcore.ETag(*g.ETag)
 	}
-	return InsertEntityResponse{
+	return UpsertEntityResponse{
 		ETag: ETag,
 	}
 }
@@ -544,16 +556,16 @@ func insertEntityFromGeneratedUpdate(g *generated.TableClientUpdateEntityRespons
 // The response type will be TableEntityMergeResponse if updateMode is Merge and TableEntityUpdateResponse if updateMode is Replace.
 // If the service returns a non-successful HTTP status code, the function returns an *azcore.ResponseError type.
 // Specify nil for options if you want to use the default options.
-func (t *Client) UpsertEntity(ctx context.Context, entity []byte, options *InsertEntityOptions) (InsertEntityResponse, error) {
+func (t *Client) UpsertEntity(ctx context.Context, entity []byte, options *UpsertEntityOptions) (UpsertEntityResponse, error) {
 	if options == nil {
-		options = &InsertEntityOptions{
+		options = &UpsertEntityOptions{
 			UpdateMode: UpdateModeMerge,
 		}
 	}
 	var mapEntity map[string]interface{}
 	err := json.Unmarshal(entity, &mapEntity)
 	if err != nil {
-		return InsertEntityResponse{}, err
+		return UpsertEntityResponse{}, err
 	}
 
 	pk := mapEntity[partitionKey]
@@ -573,6 +585,9 @@ func (t *Client) UpsertEntity(ctx context.Context, entity []byte, options *Inser
 			&generated.TableClientMergeEntityOptions{TableEntityProperties: mapEntity},
 			&generated.QueryOptions{},
 		)
+		if err != nil {
+			return UpsertEntityResponse{}, err
+		}
 		return insertEntityFromGeneratedMerge(&resp), err
 	case UpdateModeReplace:
 		resp, err := t.client.UpdateEntity(
@@ -584,12 +599,15 @@ func (t *Client) UpsertEntity(ctx context.Context, entity []byte, options *Inser
 			&generated.TableClientUpdateEntityOptions{TableEntityProperties: mapEntity},
 			&generated.QueryOptions{},
 		)
+		if err != nil {
+			return UpsertEntityResponse{}, err
+		}
 		return insertEntityFromGeneratedUpdate(&resp), err
 	}
 	if pk == "" || rk == "" {
-		return InsertEntityResponse{}, errPartitionKeyRowKeyError
+		return UpsertEntityResponse{}, errPartitionKeyRowKeyError
 	}
-	return InsertEntityResponse{}, errInvalidUpdateMode
+	return UpsertEntityResponse{}, errInvalidUpdateMode
 }
 
 // GetAccessPolicyOptions contains optional parameters for Client.GetAccessPolicy
@@ -625,6 +643,9 @@ func getAccessPolicyResponseFromGenerated(g *generated.TableClientGetAccessPolic
 // Specify nil for options if you want to use the default options.
 func (t *Client) GetAccessPolicy(ctx context.Context, options *GetAccessPolicyOptions) (GetAccessPolicyResponse, error) {
 	resp, err := t.client.GetAccessPolicy(ctx, t.name, generated.Enum4ACL, options.toGenerated())
+	if err != nil {
+		return GetAccessPolicyResponse{}, err
+	}
 	return getAccessPolicyResponseFromGenerated(&resp), err
 }
 
@@ -662,6 +683,9 @@ func (t *Client) SetAccessPolicy(ctx context.Context, options *SetAccessPolicyOp
 	response, err := t.client.SetAccessPolicy(ctx, t.name, generated.Enum4ACL, options.toGenerated())
 	if err != nil && len(options.TableACL) > 5 {
 		err = errTooManyAccessPoliciesError
+	}
+	if err != nil {
+		return SetAccessPolicyResponse{}, err
 	}
 	return setAccessPolicyResponseFromGenerated(&response), err
 }