+++ sdk/data/aztables/CHANGELOG.md
@@ -3,8 +3,10 @@
 ## 0.7.1 (Unreleased)
 
 ### Features Added
+* Added `TableErrorCode` to help recover from and understand error responses
 
 ### Breaking Changes
+* Renamed `InsertEntityResponse/Options` to `UpsertEntityResponse/Options`
 * Renamed `PossibleGeoReplicationStatusTypeValues` to `PossibleGeoReplicationStatusValues`
 
 ### Bugs Fixed